MLSample/requirements.txt
Flask==1.0.2
torchvision==0.5.0
torch==1.4.0
image-encoder==0.0.8
pika
mysql-connector
